refactor(cli): split doctor.rs into submodules + reconcile codex transport docs (#643 Phase 5 fix)
doctor.rs hit 801 lines (monolith gate 800). Split into sibling
submodules along existing section boundaries (doctor_tools,
doctor_sandbox, etc.) with no behavior change — csa doctor output is
unchanged.

Bundled with Phase 5 doc reconciliation: review pass flagged
doc-codex-transport-contract HIGH — the initial Phase 5 commit
described codex as Legacy-by-default / ACP opt-in, but runtime
defaults codex to ACP (codex_runtime.rs default_for_build) and
config validation accepts both cli/acp overrides unconditionally
(doctor surfaces missing binaries at runtime). Docs now match.

Closes #643.
