+++ README.md
@@ -110,12 +110,13 @@ See [Architecture](docs/architecture.md) for design principles and dependency gr
 | Tool | Provider | Transport | Context |
 |------|----------|-----------|---------|
 | **claude-code** | Anthropic | ACP | 200K |
-| **codex** | OpenAI | Legacy CLI (default) | 200K |
+| **codex** | OpenAI | ACP | 200K |
 | **gemini-cli** | Google | Legacy CLI | 2M |
 | **opencode** | OpenRouter | Legacy CLI | 200K |
 
-Codex ACP remains available as an opt-in path when CSA is built with
-`--features codex-acp` and `[tools.codex].transport = "acp"`.
+`claude-code` and `codex` both default to ACP runtimes today. `csa doctor`
+shows the active transport and probed binary; for codex, missing `codex-acp`
+is reported there rather than rejected during config validation.
 
 ## Documentation
